+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Clean up prompt for setting up web frameworks for Data Connect developers.
+- Fixed an issue where the prompt to create apphosting.emulator.yaml did not work with backends that are not at the project.root (#8412)